[SofaPython3] Move some examples files & remove old ones.
